Controller & Storage & All things linked together

+++ src/main/java/io/university/repository/DepartmentRepository.java
@@ -1,6 +1,6 @@
 package io.university.repository;
 
-import model.dao.Department;
+import io.university.model.dao.Department;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
@@ -11,6 +11,6 @@ import org.springframework.stereotype.Repository;
  * @since 16.02.2019
  */
 @Repository
-public interface DepartmentRepository extends JpaRepository<Department, Integer> {
+public interface DepartmentRepository extends JpaRepository<Department, Long> {
 
 }
